Embed trees into Euclidean space from Newick format

PhyloClustering.num_bipartitions — Function

num_bipartitions(n::Int64)

Function to get number of bipartitions.

Arguments

n: the number of taxa.

Output:

The number of bipartition corresponding to the number of taxa.

PhyloClustering.show_bipartitions — Function

show_bipartitions(n::Int64; start::Int64 = 0, stop::Int64=-1)

Function to show the bipartitions between certain indices of the number of taxa.

Arguments

n: the number of taxa

start(defaults to 0): the starting index.

stop(defaults to -1): the ending index.

Output

The bipartiions of trees with n taxa between start index and stop index.

Example

julia> show_bipartitions(4,stop = 4)
idx	partition
0	1 | 2 3 4 
1	2 | 1 3 4 
2	3 | 1 2 4 
3	4 | 1 2 3 
4	1 2 | 3 4 

PhyloClustering.show_bipartition — Function

show_bipartition(n::Int64, idx::Int64)

Function to show the bipartition with a given index of the number of taxa.

Arguments

n: the number of taxa.

idx: the bipartition to show.

Output

The idx-th bipartiion of trees with n taxa.

Example:

julia> show_bipartition(4, 3)
idx	partition
3	4 | 1 2 3 

PhyloClustering.get_bipartition — Function

get_bipartition(tree::HybridNetwork, n::Int64)

Function to get the bipartition format of a phylogenetic tree.

Arguments

tree: a HybridNetwork object that reperents the tree.

n: the number of taxa of the tree.

Output

A Vector of pairs shows the tree in bipartiton format.

Example

julia> get_bipartition(readTopology("(4:4.249,(1:2.457,(2:2.064,3:2.064):0.393):1.793);"), 4)
6-element Vector{Any}:
 3 => 4.249
 0 => 2.457
 1 => 2.064
 2 => 2.064
 6 => 0.393
 3 => 1.793

PhyloClustering.print_bipartition — Function

print_bipartition(trees::Vector{HybridNetwork}, n::Int64)

Function to get the bipartition format of phylogenetic trees in Matrix. Note that all taxa will be replaced by numbers. Use the function num_to_name to get a Dictionary containing the name of the taxon corresponding to the number.

Arguments

trees: a Vector of HybridNetwork objects that contains trees

n: the number of taxa of the trees

Output

A Matrix{Float64} shows the trees in bipartiton format

Example

julia> print_bipartition([readTopology("(4:4.249,(1:2.457,(2:2.064,3:2.064):0.393):1.793);"), readTopology("(4:4.308,(2:1.634,(1:1.588,3:1.588):0.046):2.674);")], 4)
2×7 Matrix{Float64}:
 2.457  2.064  2.064  6.042  0.0  0.0    0.393
 1.588  1.634  1.588  6.982  0.0  0.046  0.0

PhyloClustering.num_to_name — Function

num_to_name(tree::HybridNetwork)

Function to obtain a table that contains taxon numbers and their names

Arguments

tree: a HybridNetwork objects that contains trees

Output

A Dict{Int64, Any} shows taxon numbers and their names

Example

julia> num_to_name(readTopology("(O:3.866,(HYB:1.593,(P1:1.208,P2:1.208):0.386):2.273);"))
Dict{Int64, Any} with 4 entries:
  4 => "P2"
  2 => "O"
  3 => "P1"
  1 => "HYB"

Pre-process data before input models

PhyloClustering.standardize_tree — Function

standardize_tree(tree::AbstractMatrix{<:Real})

Function to standardize tree Matrix.

Arguments

tree: a N * B Matrix containing trees (each row is a B-dimensional tree in bipartiton format).

Output

A standardized B * N tree Matrix with a mean of about 0 and a standard deviation of about 1. This tree Matrix can be the input of model.

PhyloClustering.distance — Function

distance(tree::AbstractMatrix{<:Real})

Function to get the distance Matrix of a tree Matrix.

Arguments

tree: a B * N tree Matrix (each column of tree Matrix is a B-dimensional tree in bipartiton format).

Output

A pairwise distance Matrix that can be the input of hc_label.

Visualize results

PhyloClustering.plot_clusters — Function

plot_clusters(tree::AbstractMatrix{<:Real}, label::Vector{Int64})

Function to visualize the result of models.

Arguments

tree: a B * N tree Matrix (each column of tree Matrix is a B-dimensional tree in bipartiton format).

label: an N-length Vector containing predicted labels for each tree. People can use the output of the models.

Output

A scatter plot showing tree clusters.