K-means

K-means is a classical method for clustering or vector quantization. It produces a fixed number of clusters, each associated with a center (also known as a prototype), and each data point is assigned to a cluster with the nearest center.

From a mathematical standpoint, K-means is a coordinate descent algorithm that solves the following optimization problem:

\[\text{minimize} \ \sum_{i=1}^n \| \mathbf{x}_i - \boldsymbol{\mu}_{z_i} \|^2 \ \text{w.r.t.} \ (\boldsymbol{\mu}, z)\]

Here, $\boldsymbol{\mu}_k$ is the center of the $k$-th cluster, and $z_i$ is an index of the cluster for $i$-th point $\mathbf{x}_i$.

PhyloClustering.kmeans_label — Method

Function to get predicted labels from Yinyang K-means clustering
Input:
    tree: a B * N tree Matrix (each column of tree Matrix is a B-dimensional tree in bipartiton format)
    n: the number of clusters
    init: initialization method. K-means++ or rand
    rng: 
output:
    a Vector object with length of N containing predicted labels for each tree

Examples