Convert trees in Newick format to bipartition format

PhyloClustering.num_bipartitions — Method

Function to get number of bipartitions.
Input:
    n: the number of taxa
output:
    the number of bipartition corresponding to the number of taxa

PhyloClustering.show_bipartitions — Method

Function to show the bipartitions between certain indices of the number of taxa.
Input:
    n: the number of taxa
    start: the starting index
    stop: the ending index
output:
    the bipartiions of trees with n taxa between start index and stop index
Example:
show_bipartitions(4,stop = 4):
    idx	partition
    0	1 | 2 3 4 
    1	2 | 1 3 4 
    2	3 | 1 2 4 
    3	4 | 1 2 3 
    4	1 2 | 3 4

PhyloClustering.show_bipartition — Method

Function to show the bipartition with a given index of the number of taxa.
Input:
    n: the number of taxa
    idx: the bipartition to show
output:
    the idx-th bipartiion of trees with n taxa
    Example:
    show_bipartition(4, 3):
        idx	partition
        3	4 | 1 2 3

PhyloClustering.plot_clusters — Method

Function to visualize the result of models.
Input:
    tree: a B * N tree Matrix (each column of tree Matrix is a B-dimensional tree in bipartiton format)
    label: an N-length Vector containing predicted labels for each tree. People can use the output of the models
output:
    a scatter plot showing tree clusters

PhyloClustering.get_bipartition — Method

Function to get the bipartition format of a phylogenetic tree
Input:
    tree: a HybridNetwork object that reperents the tree
    n: the number of taxa of the tree
output:
    a Vector of pairs shows the tree in bipartiton format
Example:
    tree:
        HybridNetwork, Rooted Network
        6 edges
        7 nodes: 4 tips, 0 hybrid nodes, 3 internal tree nodes.
        tip labels: 4, 1, 2, 3
        (4:4.249,(1:2.457,(2:2.064,3:2.064):0.393):1.793);

    get_bipartition(tree, 4):
        6-element Vector{Any}:
            3 => 4.249
            0 => 2.457
            1 => 2.064
            2 => 2.064
            6 => 0.393
            3 => 1.793

PhyloClustering.print_bipartition — Method

Function to get the bipartition format of phylogenetic trees in Matrix
Note that all taxa will be replaced by numbers. Use the function num_to_name to 
get a Dictionary containing the name of the taxon corresponding to the number
Input:
    trees: a Vector of HybridNetwork objects that contains trees
    n: the number of taxa of the trees
output:
    a Matrix shows the tree in bipartiton format
Example:
    tree:
        HybridNetwork, Rooted Network
        6 edges
        7 nodes: 4 tips, 0 hybrid nodes, 3 internal tree nodes.
        tip labels: 4, 1, 2, 3
        (4:4.249,(1:2.457,(2:2.064,3:2.064):0.393):1.793);

    print_bipartition([tree], 4)
        1*7 Matrix{Float64}:
        2.457  2.064  2.064  6.042  0.0  0.0  0.393

PhyloClustering.standardize_tree — Method

Function to standardize tree Matrix.
Input:
    tree: a N * B Matrix containing trees (each row is a B-dimensional tree in bipartiton format)
output:
    a standardized B * N tree Matrix with a mean of about 0 and a standard deviation of about 1
    this tree Matrix can be the input of models

PhyloClustering.num_to_name — Method

Function to obtain a table that contains taxon numbers and their names 
Input:
    trees: a HybridNetwork objects that contains trees
output:
    a Dictionary shows taxon numbers and their names
Example:
    tree:
        HybridNetwork, Rooted Network
        6 edges
        7 nodes: 4 tips, 0 hybrid nodes, 3 internal tree nodes.
        tip labels: O, HYB, P1, P2
        (O:3.866,(HYB:1.593,(P1:1.208,P2:1.208):0.386):2.273);

    print_bipartition([tree], 4)
        Dict{Int64, Any} with 4 entries:
        4 => "P2"
        2 => "O"
        3 => "P1"
        1 => "HYB"

PhyloClustering.distance — Method

Function to get the distance Matrix of a tree Matrix.
Input:
    tree: a B * N tree Matrix
output:
    a pairwise distance Matrix that can be the input of hierarchical clustering