DBSCAN

Density-based Spatial Clustering of Applications with Noise (DBSCAN) is a data clustering algorithm that finds clusters through density-based expansion of seed points. The algorithm was proposed in:

Martin Ester, Hans-peter Kriegel, Jörg S, and Xiaowei Xu A density-based algorithm for discovering clusters in large spatial databases with noise. 1996.

Interface

The implementation of DBSCAN algorithm provided by Clustering supports the two ways of specifying clustering data:

PhyloClustering.dbscan_label — Method

Function to get predicted labels from density-based spatial clustering of applications with noise.
Input:
    tree: a B * N tree Matrix (each column of tree Matrix is a B-dimensional tree in bipartiton format) and B < N
    radius: neighborhood radius; points within this distance are considered neighbors
    min_neighbors: minimal number of neighbors required to assign a point to a cluster
    min_cluster_size: minimal number of points in a cluster
output:
    a Vector object with length of N containing predicted labels for each tree (the cluster it belongs to) 
    0 means the tree is noise